# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Kyronix Executive Presence Score API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface ErrorResponse {
  error: string;
}

export interface ScoreInput {
  /** Public LinkedIn profile URL */
  linkedin_url: string;
}

export interface EmailInput {
  email: string;
}

export interface OtpInput {
  /** 6-digit OTP code */
  code: string;
}

export type ScoreStatusStatus = typeof ScoreStatusStatus[keyof typeof ScoreStatusStatus];


export const ScoreStatusStatus = {
  pending: 'pending',
  processing: 'processing',
  complete: 'complete',
  error: 'error',
} as const;

export interface ScoreStatus {
  id: string;
  linkedin_url: string;
  status: ScoreStatusStatus;
  created_at: string;
}

export interface ScoreCategory {
  name: string;
  key: string;
  score: number;
  max_score: number;
  /** Color/tier label: elite, strong, developing, weak */
  label: string;
  /** Specific actionable improvements */
  fixes: string[];
}

export type ScoreResultStatus = typeof ScoreResultStatus[keyof typeof ScoreResultStatus];


export const ScoreResultStatus = {
  pending: 'pending',
  processing: 'processing',
  complete: 'complete',
  error: 'error',
} as const;

export interface ScoreResult {
  id: string;
  linkedin_url: string;
  status: ScoreResultStatus;
  /** @nullable */
  overall_score?: number | null;
  /**
     * Percentage of profiles this score beats
     * @nullable
     */
  percentile?: number | null;
  /**
     * Authority archetype label e.g. The Thought Leader
     * @nullable
     */
  archetype?: string | null;
  /** @nullable */
  archetype_description?: string | null;
  categories: ScoreCategory[];
  email_verified: boolean;
  /** @nullable */
  verified_email?: string | null;
  /** @nullable */
  verification_token?: string | null;
  /** @nullable */
  error_message?: string | null;
  created_at: string;
}

export interface VerificationStatus {
  success: boolean;
  message: string;
  verified: boolean;
  /**
     * JWT token to authenticate PDF download
     * @nullable
     */
  token?: string | null;
}

export type ScoreStatsTopArchetypesItem = {
  archetype: string;
  count: number;
};

export interface ScoreStats {
  total_scores: number;
  avg_overall: number;
  avg_headline: number;
  avg_visual: number;
  avg_social_proof: number;
  avg_activity: number;
  avg_content_depth: number;
  top_archetypes: ScoreStatsTopArchetypesItem[];
}