# lib/api-zod/src/generated/types/scoreCategory.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Kyronix Executive Presence Score API
 * OpenAPI spec version: 0.1.0
 */

export interface ScoreCategory {
  name: string;
  key: string;
  score: number;
  max_score: number;
  /** Color/tier label: elite, strong, developing, weak */
  label: string;
  /** Specific actionable improvements */
  fixes: string[];
}