# lib/api-zod/src/generated/types/scoreInput.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Kyronix Executive Presence Score API
 * OpenAPI spec version: 0.1.0
 */

export interface ScoreInput {
  /** Public LinkedIn profile URL */
  linkedin_url: string;
}