# lib/api-zod/src/generated/types/scoreResult.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Kyronix Executive Presence Score API
 * OpenAPI spec version: 0.1.0
 */
import type { ScoreCategory } from './scoreCategory';
import type { ScoreResultStatus } from './scoreResultStatus';

export interface ScoreResult {
  id: string;
  linkedin_url: string;
  status: ScoreResultStatus;
  /** @nullable */
  overall_score?: number | null;
  /**
     * Percentage of profiles this score beats
     * @nullable
     */
  percentile?: number | null;
  /**
     * Authority archetype label e.g. The Thought Leader
     * @nullable
     */
  archetype?: string | null;
  /** @nullable */
  archetype_description?: string | null;
  categories: ScoreCategory[];
  email_verified: boolean;
  /** @nullable */
  verified_email?: string | null;
  /** @nullable */
  verification_token?: string | null;
  /** @nullable */
  error_message?: string | null;
  created_at: Date;
}