# lib/api-zod/src/generated/types/scoreStats.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Kyronix Executive Presence Score API
 * OpenAPI spec version: 0.1.0
 */
import type { ScoreStatsTopArchetypesItem } from './scoreStatsTopArchetypesItem';

export interface ScoreStats {
  total_scores: number;
  avg_overall: number;
  avg_headline: number;
  avg_visual: number;
  avg_social_proof: number;
  avg_activity: number;
  avg_content_depth: number;
  top_archetypes: ScoreStatsTopArchetypesItem[];
}