# lib/api-zod/src/generated/types/scoreStatus.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Kyronix Executive Presence Score API
 * OpenAPI spec version: 0.1.0
 */
import type { ScoreStatusStatus } from './scoreStatusStatus';

export interface ScoreStatus {
  id: string;
  linkedin_url: string;
  status: ScoreStatusStatus;
  created_at: Date;
}