# lib/api-zod/src/generated/types/verificationStatus.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Kyronix Executive Presence Score API
 * OpenAPI spec version: 0.1.0
 */

export interface VerificationStatus {
  success: boolean;
  message: string;
  verified: boolean;
  /**
     * JWT token to authenticate PDF download
     * @nullable
     */
  token?: string | null;
}